feat: add tratarError400() method, DadosErrorValidation record to TratadorDeErrors class
The tratarError400() method is responsible for capturing the MethodArgumentNotValidException and returning a list of DadosErrorValidation object with filds campo and message and the 400 Bad Request response

The DadosErrorValidation is a DTO record that represent an error message. It has two fields: campo and message. The campo field stores the name of the field that caused the error and, the messagem field stores the error message